+++ CHANGELOG.md
@@ -115,6 +115,7 @@
 - Fixed bug in `/projects/[projectId]/dmp/[dmpId]/download` because of unused `FileIcon` [#376]
 - Fixed bug where `/template` page was continuosly refreshed when no data was returned [#351]
 - Fixed that didn't show the current title when editing a template name [#475]
+- Added the missing button to create a new template from scratch [#474]
 
 ### Chore
 - Updated and added unit tests to increase coverage in the app [#490]